# docs/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "GymCRM API",
    "description": "Multi-tenant gym SaaS. All business endpoints require Bearer JWT. gym_id is extracted from JWT — every query is automatically scoped to the authenticated gym.",
    "version": "1.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": ["http"],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter: Bearer <access_token>"
    }
  },
...
  "paths": {
...
    "/api/v1/members": {
      "post": { "tags": ["members"], "summary": "Add a new member",
        "security": [{ "BearerAuth": [] }], "consumes": ["application/json"], "produces": ["application/json"],
        "parameters": [{ "in": "body", "name": "body", "required": true, "schema": { "$ref": "#/definitions/CreateMemberRequest" } }],
        "responses": { "201": { "description": "Created", "schema": { "$ref": "#/definitions/MemberResponse" } }, "409": { "description": "Phone exists" }, "422": { "description": "Validation error" } } },
      "get": { "tags": ["members"], "summary": "List members",
        "security": [{ "BearerAuth": [] }], "produces": ["application/json"],
        "parameters": [
          { "in": "query", "name": "page", "type": "integer" },
          { "in": "query", "name": "per_page", "type": "integer" },
          { "in": "query", "name": "status", "type": "string", "description": "active | expired | inactive | churned" },
          { "in": "query", "name": "search", "type": "string" }
        ],
        "responses": { "200": { "description": "OK", "schema": { "$ref": "#/definitions/MemberListResponse" } } } }
    },
    "/api/v1/members/search": {
      "get": { "tags": ["members"], "summary": "Search members",
        "security": [{ "BearerAuth": [] }], "produces": ["application/json"],
        "parameters": [
          { "in": "query", "name": "q", "type": "string", "required": true },
          { "in": "query", "name": "page", "type": "integer" },
          { "in": "query", "name": "per_page", "type": "integer" }
        ],
        "responses": { "200": { "description": "OK", "schema": { "$ref": "#/definitions/MemberListResponse" } } } }
    },
    "/api/v1/members/expiring": {
      "get": { "tags": ["members"], "summary": "Get expiring members",
        "description": "Members expiring within N days. Sorted by expiry date ascending.",
        "security": [{ "BearerAuth": [] }], "produces": ["application/json"],
        "parameters": [
          { "in": "query", "name": "days", "type": "integer", "description": "Days ahead (default: 7, max: 90)" },
          { "in": "query", "name": "page", "type": "integer" },
          { "in": "query", "name": "per_page", "type": "integer" }
        ],
        "responses": { "200": { "description": "OK", "schema": { "$ref": "#/definitions/MemberListResponse" } } } }
    },
    "/api/v1/members/{id}": {
      "get": { "tags": ["members"], "summary": "Get member by ID",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "in": "path", "name": "id", "type": "integer", "required": true }],
        "responses": { "200": { "description": "OK", "schema": { "$ref": "#/definitions/MemberResponse" } }, "404": { "description": "Not found" } } },
      "put": { "tags": ["members"], "summary": "Update member",
        "security": [{ "BearerAuth": [] }], "consumes": ["application/json"],
        "parameters": [
          { "in": "path", "name": "id", "type": "integer", "required": true },
          { "in": "body", "name": "body", "required": true, "schema": { "$ref": "#/definitions/UpdateMemberRequest" } }
        ],
        "responses": { "200": { "description": "OK", "schema": { "$ref": "#/definitions/MemberResponse" } }, "404": { "description": "Not found" } } },
      "delete": { "tags": ["members"], "summary": "Delete member (soft delete)",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "in": "path", "name": "id", "type": "integer", "required": true }],
        "responses": { "200": { "description": "Deleted" }, "404": { "description": "Not found" } } }
    },
...
  },
  "definitions": {
...
    "CreateMemberRequest": {
      "type": "object", "required": ["first_name", "last_name", "phone"],
      "properties": {
        "first_name": { "type": "string", "example": "Harpreet" },
        "last_name": { "type": "string", "example": "Kaur" },
        "phone": { "type": "string", "example": "9876500001" },
        "email": { "type": "string" }, "gender": { "type": "string", "enum": ["male", "female", "other"] },
        "date_of_birth": { "type": "string", "example": "1995-06-15" },
        "address": { "type": "string" },
        "membership_plan_id": { "type": "integer" },
        "start_date": { "type": "string", "example": "2026-06-01" },
        "expiry_date": { "type": "string", "example": "2026-06-30" },
        "notes": { "type": "string" }
      }
    },
    "UpdateMemberRequest": {
      "type": "object",
      "properties": {
        "first_name": { "type": "string" }, "last_name": { "type": "string" },
        "phone": { "type": "string" }, "email": { "type": "string" },
        "gender": { "type": "string", "enum": ["male", "female", "other"] },
        "address": { "type": "string" }, "membership_plan_id": { "type": "integer" },
        "start_date": { "type": "string" }, "expiry_date": { "type": "string" },
        "status": { "type": "string", "enum": ["active", "expired", "inactive", "churned"] },
        "notes": { "type": "string" }
      }
    },
    "MemberResponse": {
      "type": "object",
      "properties": {
        "id": { "type": "integer" }, "gym_id": { "type": "integer" },
        "first_name": { "type": "string" }, "last_name": { "type": "string" }, "full_name": { "type": "string" },
        "phone": { "type": "string" }, "email": { "type": "string" }, "gender": { "type": "string" },
        "date_of_birth": { "type": "string" }, "address": { "type": "string" },
        "membership_plan_id": { "type": "integer" },
        "start_date": { "type": "string" }, "expiry_date": { "type": "string" },
        "status": { "type": "string" }, "notes": { "type": "string" },
        "created_at": { "type": "string" }, "updated_at": { "type": "string" }
      }
    },
    "MemberListResponse": {
      "type": "object",
      "properties": {
        "success": { "type": "boolean" },
        "data": { "type": "object", "properties": {
          "members": { "type": "array", "items": { "$ref": "#/definitions/MemberResponse" } }
        }},
        "meta": { "type": "object", "properties": {
          "page": { "type": "integer" }, "per_page": { "type": "integer" },
          "total": { "type": "integer" }, "total_pages": { "type": "integer" }
        }}
      }
    },
...
  }
}